requirements.txt
aiofiles==24.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
html5tagger==1.3.0
httptools==0.6.4
idna==3.10
multidict==6.1.0
peewee==3.17.7
pycparser==2.22
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
ruff==0.7.3
sanic==24.6.0
sanic-ext==23.12.0
sanic-routing==23.12.0
setuptools==75.4.0
tracerite==1.1.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvloop==0.21.0
websockets==14.0
